refactor: oauth2 로그인에 필요한 객체 필드 및 메서드 정리
#35
#5
