+++ src/main/java/MusicPlatform/domain/oauth2/entity/GoogleUser.java
@@ -4,9 +4,7 @@ import MusicPlatform.domain.oauth2.adapter.AuthenticationAdapter;
 import org.springframework.security.oauth2.client.registration.ClientRegistration;
 import org.springframework.security.oauth2.core.user.OAuth2User;
 
-public class GoogleUser extends OAuth2ProviderUser implements AuthenticationAdapter {
-
-    private String uuid;
+public class GoogleUser extends OAuth2ProviderUser {
 
     public GoogleUser(OAuth2User oAuth2User, ClientRegistration clientRegistration) {
         super(oAuth2User.getAttributes(), oAuth2User, clientRegistration);
@@ -25,15 +23,5 @@ public class GoogleUser extends OAuth2ProviderUser implements AuthenticationAdap
     @Override
     public String getPicture() {
         return getAttributes().get("picture").toString();
-    }
-
-    @Override
-    public void setUuid(String uuid) {
-        this.uuid = uuid;
-    }
-
-    @Override
-    public String getUuid() {
-        return this.uuid;
     }
 }
