+++ src/main/java/MusicPlatform/domain/oauth2/entity/OAuth2ProviderUser.java
@@ -1,5 +1,6 @@
 package MusicPlatform.domain.oauth2.entity;
 
+import MusicPlatform.domain.oauth2.adapter.AuthenticationAdapter;
 import java.util.List;
 import java.util.Map;
 import java.util.UUID;
@@ -11,11 +12,12 @@ import org.springframework.security.oauth2.core.user.OAuth2User;
 /**
  * 모든 서비스 제공자가 동일하게 제공하는 속성들
  */
-public abstract class OAuth2ProviderUser implements ProviderUser {
+public abstract class OAuth2ProviderUser implements ProviderUser, AuthenticationAdapter {
 
     private final Map<String, Object> attributes;
     private final OAuth2User oAuth2User;
     private final ClientRegistration clientRegistration;
+    private String uuid;
 
     public OAuth2ProviderUser(Map<String, Object> attributes, OAuth2User oAuth2User,
                               ClientRegistration clientRegistration) {
@@ -49,5 +51,15 @@ public abstract class OAuth2ProviderUser implements ProviderUser {
     @Override
     public Map<String, Object> getAttributes() {
         return attributes;
+    }
+
+    @Override
+    public void setUuid(String uuid) {
+        this.uuid = uuid;
+    }
+
+    @Override
+    public String getUuid() {
+        return this.uuid;
     }
 }
